refactor(config): extract ConfigError and guards into separate modules
- ConfigError moved to src/errors.ts (room for future error classes).
- check() and isPlainObject() moved to src/guards.ts (generic predicates,
  reusable by future validators).
- check() is now generic: takes an Error instance instead of a message,
  so it can be paired with any error class.

src/config.ts: 144 → 132 lines.
