+++ src/__tests__/config.test.ts
@@ -2,7 +2,8 @@ import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { ConfigError, DEFAULT_CONFIG, loadConfig, validateAndMerge } from "../config.ts";
+import { DEFAULT_CONFIG, loadConfig, validateAndMerge } from "../config.ts";
+import { ConfigError } from "../errors.ts";
 
 let workDir: string;
 let homeDir: string;
